Don't allow to remove account if there are shared documents

+++ app/controllers/editor/users_controller.rb
@@ -13,9 +13,14 @@ class Editor::UsersController < EditorController
   end
 
   def destroy
-    current_user.collaborations.delete_all
-    current_user.delete
-    redirect_to root_path, alert: "Your account has been removed"
+    if current_user.own_shared_documents?
+      redirect_to editor_profile_path, alert: "You can't remove the account until you pass the ownership of shared documents."
+    else
+      current_user.collaborations.delete_all
+      current_user.bookmarks.delete_all
+      current_user.delete
+      redirect_to root_path, notice: "Your account has been removed"
+    end
   end
 
   private
